Restructure meshsim to be more readable
