+++ .gitignore
@@ -1,1 +1,2 @@
 /pg-mesh-data
+*.pyc
