+++ meshsim/config.py
@@ -1,0 +1,6 @@
+config = None
+
+
+def set_config(_config):
+    global config
+    config = _config
